+++ brahmand/src/clickhouse_query_generator/errors.rs
@@ -48,4 +48,6 @@ pub enum ClickhouseQueryGeneratorError {
         "Unknow 'To' table found in relationship defination. Make sure to create nodes first before creating relationship."
     )]
     UnknownToTableInRel,
+    #[error("Schema error: {0}")]
+    SchemaError(String),
 }