+++ brahmand/src/query_planner/analyzer/filter_tagging.rs
@@ -85,6 +85,7 @@ impl AnalyzerPass for FilterTagging {
                 }
                 union.rebuild_or_clone(inputs_tf, logical_plan.clone())
             }
+            LogicalPlan::PageRank(_) => Transformed::No(logical_plan.clone()),
         };
         Ok(transformed_plan)
     }