+++ brahmand/src/query_planner/analyzer/graph_join_inference.rs
@@ -130,6 +130,7 @@ impl GraphJoinInference {
                 }
                 union.rebuild_or_clone(inputs_tf, logical_plan.clone())
             },
+            LogicalPlan::PageRank(_) => Transformed::No(logical_plan.clone()),
             LogicalPlan::ViewScan(_) => Transformed::No(logical_plan.clone())
         };
         Ok(transformed_plan)
@@ -243,6 +244,7 @@ impl GraphJoinInference {
                 }
                 Ok(())
             }
+            LogicalPlan::PageRank(_) => Ok(()),
         }
     }