+++ brahmand/src/query_planner/analyzer/group_by_building.rs
@@ -92,6 +92,7 @@ impl AnalyzerPass for GroupByBuilding {
                 }
                 union.rebuild_or_clone(inputs_tf, logical_plan.clone())
             }
+            LogicalPlan::PageRank(_) => Transformed::No(logical_plan.clone()),
         };
         Ok(transformed_plan)
     }