+++ brahmand/src/query_planner/analyzer/plan_sanitization.rs
@@ -113,6 +113,7 @@ impl PlanSanitization {
                 }
                 union.rebuild_or_clone(inputs_tf, logical_plan.clone())
             }
+            LogicalPlan::PageRank(_) => Transformed::No(logical_plan.clone()),
         };
         Ok(transformed_plan)
     }