+++ brahmand/src/query_planner/analyzer/projection_tagging.rs
@@ -140,6 +140,7 @@ impl AnalyzerPass for ProjectionTagging {
                 }
                 union.rebuild_or_clone(inputs_tf, logical_plan.clone())
             }
+            LogicalPlan::PageRank(_) => Transformed::No(logical_plan.clone()),
             LogicalPlan::ViewScan(_view_scan) => Transformed::No(logical_plan.clone()),
         };
         Ok(transformed_plan)