+++ brahmand/src/query_planner/analyzer/schema_inference.rs
@@ -115,6 +115,7 @@ impl SchemaInference {
                 }
                 union.rebuild_or_clone(inputs_tf, logical_plan.clone())
             }
+            LogicalPlan::PageRank(_) => Transformed::No(logical_plan.clone()),
             LogicalPlan::ViewScan(_) => Transformed::No(logical_plan.clone()),
         };
         Ok(transformed_plan)
@@ -243,6 +244,7 @@ impl SchemaInference {
                 }
                 Ok(())
             }
+            LogicalPlan::PageRank(_) => Ok(()),
             LogicalPlan::ViewScan(_) => Ok(()),
         }
     }