+++ brahmand/src/query_planner/optimizer/filter_into_graph_rel.rs
@@ -301,6 +301,7 @@ impl OptimizerPass for FilterIntoGraphRel {
                 }
                 union.rebuild_or_clone(inputs_tf, logical_plan.clone())
             }
+            LogicalPlan::PageRank(_) => Transformed::No(logical_plan.clone()),
         };
         Ok(transformed_plan)
     }