+++ brahmand/src/query_planner/optimizer/projection_push_down.rs
@@ -91,6 +91,7 @@ impl OptimizerPass for ProjectionPushDown {
                 }
                 union.rebuild_or_clone(inputs_tf, logical_plan.clone())
             }
+            LogicalPlan::PageRank(_) => Transformed::No(logical_plan.clone()),
             LogicalPlan::ViewScan(_) => Transformed::No(logical_plan.clone())
         };
         Ok(transformed_plan)